requirements.txt
APScheduler==3.11.2
blinker==1.9.0
branca==0.8.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
folium==0.20.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pyproj==3.7.2
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
shapely==2.1.2
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
xyzservices==2025.11.0
